+++ packages/service-library/src/servicelib/long_running_tasks/_constants.py
@@ -1,0 +1,5 @@
+from typing import Final
+
+MINUTE: Final[int] = 60  # in secs
+HOUR: Final[int] = 60 * MINUTE  # in secs
+DEFAULT_POLL_INTERVAL_S: Final[float] = 1
